+++ EpsilonScript/AST/NegateNode.cs
@@ -44,5 +44,10 @@ namespace EpsilonScript.AST
       _childNode = _childNode.Optimize();
       return this;
     }
+
+    public override void ConfigureNoAlloc()
+    {
+      _childNode?.ConfigureNoAlloc();
+    }
   }
 }