+++ EpsilonScript/AST/SequenceNode.cs
@@ -92,5 +92,11 @@ namespace EpsilonScript.AST
       _rightNode = _rightNode.Optimize();
       return this;
     }
+
+    public override void ConfigureNoAlloc()
+    {
+      _leftNode?.ConfigureNoAlloc();
+      _rightNode?.ConfigureNoAlloc();
+    }
   }
 }